requirements.txt
# core deep learning libraries
torch==2.0.1
torchvision==0.15.2
pytorch-lightning==1.9.5

# Transformer models
transformers==4.30.0
sentence-transformers==2.2.2

# data processing
numpy==1.24.3
pandas==2.0.3
scipy==1.10.1

# machine learning tools
scikit-learn==1.3.2

# visible interface
matplotlib==3.7.2
seaborn==0.12.2

# NLP
nltk==3.8.1

# medical NLP (uncomment if needed)
# umls==0.1.0

# toolkits
tqdm==4.65.0
pyyaml==6.0
pillow==9.5.0
opencv-python==4.8.0.74